requirements.txt
flask
pandas
scikit-learn
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
numpy
torch